Moved `autofocus` from SBIG `Camera` to `AbstractCamera` base class
